+++ src/markitdown/converters/__init__.py
@@ -10,6 +10,11 @@ from ._wikipedia_converter import WikipediaConverter
 from ._youtube_converter import YouTubeConverter
 from ._ipynb_converter import IpynbConverter
 from ._bing_serp_converter import BingSerpConverter
+from ._pdf_converter import PdfConverter
+from ._docx_converter import DocxConverter
+from ._xlsx_converter import XlsxConverter, XlsConverter
+from ._pptx_converter import PptxConverter
+from ._image_converter import ImageConverter
 
 __all__ = [
     "DocumentConverter",
@@ -21,4 +26,10 @@ __all__ = [
     "YouTubeConverter",
     "IpynbConverter",
     "BingSerpConverter",
+    "PdfConverter",
+    "DocxConverter",
+    "XlsxConverter",
+    "XlsConverter",
+    "PptxConverter",
+    "ImageConverter",
 ]
